+++ src/server/index.ts
@@ -1,0 +1,36 @@
+import express from "express";
+import { PrismaClient } from "@prisma/client";
+import cors from "cors";
+
+const app = express();
+const prisma = new PrismaClient();
+
+app.use(cors());
+app.use(express.json());
+
+app.get("/api/users", async (req, res) => {
+  try {
+    const users = await prisma.user.findMany();
+    res.json(users);
+  } catch (error) {
+    res.status(500).json({ error: "Failed to fetch users" });
+  }
+});
+
+app.get("/api/posts", async (req, res) => {
+  try {
+    const posts = await prisma.post.findMany({
+      include: {
+        author: true,
+      },
+    });
+    res.json(posts);
+  } catch (error) {
+    res.status(500).json({ error: "Failed to fetch posts" });
+  }
+});
+
+const PORT = process.env.PORT || 3001;
+app.listen(PORT, () => {
+  console.log(`Server running on port ${PORT}`);
+});
