+++ vercel.json
@@ -2,7 +2,7 @@
   "version": 2,
   "builds": [
     {
-      "src": "api/index.js",
+      "src": "src/server/index.ts",
       "use": "@vercel/node"
     },
     {
@@ -14,11 +14,11 @@
   "routes": [
     {
       "src": "/api/(.*)",
-      "dest": "/api/index.js"
+      "dest": "src/server/index.ts"
     },
     {
       "src": "/(.*)",
-      "dest": "/$1"
+      "dest": "build/$1"
     }
   ]
 }
